requirements.txt

# streamlit>=1.25.0
# python-dotenv>=1.0.0
# pypdf>=4.2.0

# # Modern sentence-transformers bundle (no deprecated stuff)
# sentence-transformers>=3.0.0

# # Prebuilt FAISS wheels for macOS ARM
# faiss-cpu==1.7.4

# # Latest transformers + tokenizers (Python 3.13-compatible)
# transformers>=4.44.0
# tokenizers>=0.19.0

# # Latest PyTorch nightly supports Python 3.13 + macOS ARM
# torch==2.5.0
# numpy>=1.26.0

streamlit>=1.26.0
python-dotenv>=1.0.0
pypdf>=4.2.0
sentence-transformers>=3.0.0
chromadb>=0.4.22
numpy>=1.26.0
requests>=2.31.0
